--- idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmResult/index.ts
@@ -1,2 +1,0 @@
-export * from "./components/AlgorithmSubmitResult/AlgorithmSubmitResult";
-export * from "./components/AlgorithmTestResult/AlgorithmTestResult";
