+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmSubmitResult/AlgorithmSubmitResult.styles.tsx
