+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmTestResult/AlgorithmTestResult.styles.tsx
